+++ authors.yaml
@@ -52,3 +52,7 @@ jamesle:
   name: "James Le"
   website: "https://x.com/le_james94"
   avatar: "/blog/img/author/james-le.jpg"
+carlosesteban:
+  name: "Carlos Esteban"
+  website: "https://www.linkedin.com/in/cmesteban/"
+  avatar: "/blog/img/author/carlos-esteban.jpg"
